update print and layour rs for modularization
